Modify Double UAV out of range problem
Reformulated the problem to have a fixed end time, with the aim to maximize the in-range time in this period.

Adjusted the objective formulation with Alvaro and Ju's proposal, but incorporated the LogSumExp idea you have found.

Made simple modifications to the battery model to incorporate wireless charging (and made the dynamics less stiff).

Reverted the target shape to the one Ju uses.